input_chunk: split groups of records if incoming append raw data is large
As of now we do not enforce the size of the incoming data when the function
flb_input_chunk_append_raw() is invoked by input plugins. This patch aims
to remediate that situation.

If the incoming msgpack size is greater than our chunk size limit (2MB), it
will iterate the records and try to ingest them by groups up to 2MB. This
has a small performance penalty but it helps to ensure chunks sizes will
remain under the limit. But 'note; that if a single record is bigger than
our chunk size it will pass, that's fine and expected.